removed separate routers and moved models folder out. clarified README
