+++ frontend/src/pages/Home.jsx
@@ -2,5 +2,5 @@ export default function HomePage() {
   return <>
     <h1>Home</h1>
     <p>Put something interesting here!</p>
-  </>;
+  </>
 }
