+++ server/db/seeds/init.js
@@ -9,6 +9,7 @@ exports.seed = async (knex) => {
 
   await knex.raw('ALTER SEQUENCE users_id_seq RESTART WITH 1');
 
+  // User.create(username, password)
   await User.create('cool_cat', '1234');
   await User.create('l33t-guy', '1234');
   await User.create('wowow', '1234');
